+++ configure.js
@@ -1,5 +1,11 @@
 const { setWeb3Provider } = require('./src/config/web3');
-const { setEnvironment } = require('./src/config/environment');
+const { setSingletonsConfig } = require('./src/config/singletons');
+
+const { deprecate } = require('util');
+
+const setEnvironment = deprecate((environment) => setSingletonsConfig({ abstraction: environment }),
+  'The \'environment\' configuration option is deprecated, use \'singletons.abstraction\' instead.'
+);
 
 let configLoaded = false;
 
@@ -18,7 +24,7 @@ function configure (config) {
 
 function defaultConfigure () {
   setWeb3Provider.default();
-  setEnvironment.default();
+  setSingletonsConfig.default();
 }
 
 function customConfigure (config) {
@@ -29,7 +35,9 @@ function customConfigure (config) {
     setWeb3Provider(config.provider);
   }
 
-  if ('environment' in config) {
+  if ('singletons' in config) {
+    setSingletonsConfig(config.singletons);
+  } else if ('environment' in config) {
     setEnvironment(config.environment);
   }
 }