+++ test-integration/truffle-migration/migrations/1_initial_migration.js
@@ -4,7 +4,7 @@ module.exports = function(deployer) {
   deployer.deploy(Migrations);
 
   try {
-    require('@openzeppelin/test-helpers/configure')({ provider: web3.currentProvider, environment: 'truffle' });
+    require('@openzeppelin/test-helpers/configure')({ provider: web3.currentProvider, singletons: { abstraction: 'truffle' } });
 
     console.error('Successfully configured Web3 instance');
   } catch (e) {