+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.5.4 (unreleased)
  * Fixed some RPC calls not having an `id` field. ([#92](https://github.com/OpenZeppelin/openzeppelin-test-helpers/pull/92))
+ * Added `singletons` configuration and renamed the `environment` option to `singletons.abstraction`. ([#96](https://github.com/OpenZeppelin/openzeppelin-test-helpers/pull/96))
 
 ## 0.5.3 (2019-10-16)
  * Fixed a bug in the `chai-bn` setup. ([#85](https://github.com/OpenZeppelin/openzeppelin-test-helpers/pull/85))